move '_core.Event''s unit tests to separated file
